Use image.Set to recolor the image, output the result with jpeg.Encode
This commit replaces the code for counting light and dark pixels with
code that recolors pixels, applying them to an image.RGBA that is then
returned from the recolor function.

Additionally, before outputting the result of the recolor function to
the recolored.jpg file, we have a dialogue asking whether the user
wants to replace the recolored.jpg file if it already exists.